+++ example-blank/spec/handler_spec.rb
@@ -5,7 +5,7 @@ describe 'handler(event)' do
   let(:event) {SpecHelper::Event.new}
 
   it 'should return Hash, String or Array' do
-    body = handler(event).body
-    expect([String, Hash, Array].include? body.class).to be true
+    body = handler(event).call.body
+    expect([String, Hash, Array, ].include? body.class).to be true
   end
 end