+++ example/spec/spec_helper.rb
@@ -1,3 +1,3 @@
-require 'faastruby-rpc'
+require 'faastruby-rpc/test_helper'
 require 'helpers/faastruby'
 include FaaStRuby