+++ lib/faastruby/cli/commands/function.rb
@@ -22,7 +22,7 @@ module FaaStRuby
 end
 
 require 'faastruby/cli/commands/function/build'
-require 'faastruby/cli/commands/function/deploy'
+require 'faastruby/cli/commands/function/deploy_to'
 require 'faastruby/cli/commands/function/new'
 require 'faastruby/cli/commands/function/remove_from'
 require 'faastruby/cli/commands/function/test'