+++ lib/faastruby/cli/commands/workspace.rb
@@ -10,3 +10,4 @@ end
 require 'faastruby/cli/commands/workspace/create'
 require 'faastruby/cli/commands/workspace/destroy'
 require 'faastruby/cli/commands/workspace/list'
+require 'faastruby/cli/commands/workspace/deploy'